rota-backend/requirements.txt
# Şu an için herhangi bir harici kütüphane gerekmiyor.
# İlerleyen aşamalarda gerekli kütüphaneler buraya eklenecek. 

flask==2.0.1
flask-cors==3.0.10
geopy==2.4.1
shapely==2.0.2
googlemaps==4.10.0
python-dotenv==1.0.0
gunicorn==20.1.0 